backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
sqlmodel==0.0.22
alembic==1.14.0
openai==1.54.0
httpx==0.27.0
firebase-admin==6.5.0

